refactor: suggestions into smaller components
